Redistribute torch marshall gap tests into feature-aligned files (#782)
Move tests from test_torch_marshall_gaps.py into test_torchintegration.py
(factory paths, properties, error guards, nanobind bindings) and
test_torch_autograd_workflows.py (backward pass DiffPair output).
Broaden parametrization from CUDA-only to DEVICE_TYPES and switch
dispatch tests to get_torch_device() for Vulkan interop support.

Made-with: Cursor
